requirements — копия.txt
annotated-types==0.7.0
anyio==4.5.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dacktool==0.0.7
dbstream==0.1.28
distro==1.9.0
docx==0.2.4
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastapi==0.115.12
git-filter-repo==2.47.0
google-api-core==2.25.0rc0
google-api-python-client==1.7.11
google-auth==2.39.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==1.2.2
google-cloud==0.34.0
google-cloud-bigquery==3.30.0
google-cloud-bigquery-storage==2.31.0
google-cloud-core==2.4.3
google-cloud-secret-manager==2.7.2
google-crc32c==1.6.0.dev2
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googleauthentication==0.0.18
grpc-google-iam-v1==0.12.7
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.4.5
jiter==0.9.0
libcst==1.1.0
lxml==5.3.2
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.71.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.1
pillow==10.4.0
proto-plus==1.26.1
protobuf==4.25.7
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools>=65.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
tiktoken==0.7.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==3.0.1
urllib3==2.2.3
uvicorn==0.33.0
zipp==3.20.2